Conflicts:
	README.rst
	autotest

+++ jenkins/run_tests.sh
@@ -45,7 +45,7 @@ python manage.py loaddata auth_data.json range_types.json --traceback
 python manage.py eoxs_rangetype_load -i autotest_jenkins/data/meris/meris_range_type_definition.json --traceback
 python manage.py eoxs_rangetype_load -i autotest_jenkins/data/asar/asar_range_type_definition.json --traceback
 python manage.py eoxs_rangetype_list --traceback
-python manage.py eoxs_rangetype_list --json --traceback # dump all range-types as JSON 
+python manage.py eoxs_rangetype_list --json --traceback # dump all range-types as JSON
 python manage.py eoxs_collection_create -i MER_FRS_1P_reduced --traceback
 python manage.py eoxs_dataset_register -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.tif -m autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.xml -r MERIS_uint16 --visible --traceback
 python manage.py eoxs_dataset_register -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.tif -m autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.xml -r MERIS_uint16 --visible --traceback
@@ -75,8 +75,10 @@ xmllint --format tmp > tmp3
 curl -sS -o tmp "http://localhost:8000/ows?service=WCS&version=2.0.1&request=DescribeEOCoverageSet&eoId=MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed"
 xmllint --format tmp > tmp4
 
-# Restart development server otherwise the GetCoverage requests hangs forever
-kill `ps --ppid $PID -o pid=`
+# Restart development server otherwise the GetCoverage request hangs forever
+if ps --ppid $PID -o pid=; then
+    kill `ps --ppid $PID -o pid=`
+fi
 python manage.py runserver 1>/dev/null 2>&1 &
 sleep 3
 PID=$!
@@ -95,7 +97,9 @@ if [ $OS != "Ubuntu" ]; then
 fi
 
 rm tmp tmp1 tmp2 tmp3 tmp4
-kill `ps --ppid $PID -o pid=`
+if ps --ppid $PID -o pid=; then
+    kill `ps --ppid $PID -o pid=`
+fi
 
 python manage.py eoxs_collection_unlink --collection MER_FRS_1P_reduced --remove mosaic_MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_RGB_reduced mosaic_MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_RGB_reduced mosaic_MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_RGB_reduced --traceback
 python manage.py eoxs_dataset_deregister mosaic_MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_RGB_reduced mosaic_MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_RGB_reduced mosaic_MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_RGB_reduced --traceback
@@ -110,6 +114,25 @@ python manage.py eoxs_id_check --traceback nonexistent_XYZ || echo "Test of 'eox
 #python manage.py eoxs_collection_create -i test_sync -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ autotest_jenkins/data/meris/mosaic_MER_FRS_1P_reduced_RGB/ -p "*.tif" --traceback
 #python manage.py eoxs_synchronize -a --traceback
 python manage.py eoxs_id_list --traceback
+
+# deregister all of MER_FRS_1P_reduced
+python manage.py eoxs_dataset_deregister MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_uint16_reduced_compressed
+
+# test batch registration
+echo "\"data\",\"metadata\",\"range_type_name\",\"visible\",\"collection\"
+\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.tif\",\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.xml\",\"MERIS_uint16\",TRUE,\"MER_FRS_1P_reduced\"
+\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.tif\",\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.xml\",\"MERIS_uint16\",TRUE,\"MER_FRS_1P_reduced\"
+\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_uint16_reduced_compressed.tif\",\"autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_uint16_reduced_compressed.xml\",\"MERIS_uint16\",TRUE,\"MER_FRS_1P_reduced\"
+" > batch.csv
+
+python manage.py eoxs_dataset_register_batch batch.csv --traceback
+
+rm batch.csv
+
+# test --replace flag
+python manage.py eoxs_dataset_register -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.tif -m autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060816_090929_000001972050_00222_23322_0058_uint16_reduced_compressed.xml -r MERIS_uint16 --visible --replace --traceback
+python manage.py eoxs_dataset_register -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.tif -m autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060822_092058_000001972050_00308_23408_0077_uint16_reduced_compressed.xml -r MERIS_uint16 --visible --replace --traceback
+python manage.py eoxs_dataset_register -d autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_uint16_reduced_compressed.tif -m autotest_jenkins/data/meris/MER_FRS_1P_reduced/ENVISAT-MER_FRS_1PNPDE20060830_100949_000001972050_00423_23523_0079_uint16_reduced_compressed.xml -r MERIS_uint16 --visible --replace --traceback
 
 # Run Selenium
 echo "**> running Selenium tests ..."
